smac/pyproject.toml
[tool.poetry]
name = "smac"
version = "0.1.0"
description = ""
authors = ["Marco Incerti <marco.incerti@satispay.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
numpy = ">=2.0.2,<3.0.0"
pomegranate = "^1.1.1"
pandas = "^2.2.3"
openpyxl = "^3.1.5"
xlrd = "^2.0.1"
pyro-ppl = "^1.9.1"
matplotlib = "^3.9.2"
scipy = "^1.14.1"
opencv-python = "^4.10.0.84"
hmmlearn = "^0.3.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pandas = "^2.2.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
